refactor: tidy test case
